refactor: clean types and reorder imports
