+++ src/data/questions.ts
@@ -1,20 +1,4 @@
-// Type de question
-export type QuestionType =
-  | {
-      type: 'text';
-      label: string;
-    }
-  | {
-      type: 'radio';
-      label: string;
-      options: string[];
-    }
-  | {
-      type: 'checkbox';
-      label: string;
-      options: string[];
-      max?: number;
-    };
+import { type QuestionType } from '@/types/question';
 
 export const questions: QuestionType[] = [
   {
