+++ src/types/answers.d.ts
@@ -1,0 +1,1 @@
+export type AnswersType = Record<number, string[]>;
