+++ src/types/question.d.ts
@@ -1,0 +1,16 @@
+export type QuestionType =
+  | {
+      type: 'text';
+      label: string;
+    }
+  | {
+      type: 'radio';
+      label: string;
+      options: string[];
+    }
+  | {
+      type: 'checkbox';
+      label: string;
+      options: string[];
+      max?: number;
+    };
